Split internal curve models into a private submodule
